+++ x/observer/keeper/vote_inbound_test.go
@@ -24,7 +24,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err := k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			sample.AccAddress(),
 			"index",
@@ -46,7 +46,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err := k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			sample.AccAddress(),
 			"index",
@@ -68,7 +68,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err = k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			sample.AccAddress(),
 			"index",
@@ -97,7 +97,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err := k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			sample.AccAddress(),
 			"index",
@@ -134,7 +134,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err := k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			observer,
 			"index",
@@ -151,7 +151,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 					IsSupported: true,
 				},
 				{
-					ChainId:     chains.ZetaPrivnetChain().ChainId,
+					ChainId:     chains.ZetaPrivnetChain.ChainId,
 					IsSupported: false,
 				},
 			},
@@ -162,7 +162,7 @@ func TestKeeper_VoteOnInboundBallot(t *testing.T) {
 		_, _, err = k.VoteOnInboundBallot(
 			ctx,
 			getValidEthChainIDWithIndex(t, 0),
-			chains.ZetaPrivnetChain().ChainId,
+			chains.ZetaPrivnetChain.ChainId,
 			coin.CoinType_ERC20,
 			observer,
 			"index",