+++ x/observer/types/observer.pb.go
@@ -87,6 +87,7 @@ func (ObserverUpdateReason) EnumDescriptor() ([]byte, []int) {
 	return fileDescriptor_3004233a4a5969ce, []int{1}
 }
 
+// Deprecated: Use ObserverSet instead to get the list of observers
 type ObserverMapper struct {
 	Index         string        `protobuf:"bytes,1,opt,name=index,proto3" json:"index,omitempty"`
 	ObserverChain *chains.Chain `protobuf:"bytes,2,opt,name=observer_chain,json=observerChain,proto3" json:"observer_chain,omitempty"`